pyproject.toml
[project]
name = "course-management-platform"
version = "0.1.0"
description = "A Django-based course management platform"
requires-python = ">=3.13"

dependencies = [
    "django",
    "django-allauth[slack]",
    "requests",
    "pyjwt",
    "cryptography",
    "dj-database-url",
    "whitenoise",
    "psycopg2-binary",
    "gunicorn",
    "python-json-logger",
    "django-unfold",
    "django-loginas",
]

[dependency-groups]
dev = [
    "ipython",
    "pytest-django",
    "pytest",
    "jupyter",
    "pandas",
    "tqdm",
    "playwright>=1.58.0",
]


[tool.ruff]
line-length = 72

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "course_management.settings"
python_files = ["test_*.py"]

testpaths = [
    "courses/tests",
    "data/tests",
    "cadmin/tests",
]

[tool.uv]
exclude-newer = "7 days"
